警告対策。

+++ NVEncCore/rgy_util.cpp
@@ -1181,14 +1181,14 @@ unsigned short float2half(float value) {
             // underflow
             if (newexp >= -10) {
                 // denormal half-precision
-                unsigned short sig = (significand | 0x800000) >> (14 - newexp);
+                unsigned short sig = (unsigned short)((significand | 0x800000) >> (14 - newexp));
                 fp16 = (sign << 15) | (0x00 << 10) | sig;
             } else {
                 // underflow
                 fp16 = (sign << 15) | (0x00 << 10) | 0x00;
             }
         } else {
-            fp16 = (sign << 15) | (newexp << 10) | (significand >> 13);
+            fp16 = (unsigned short)((sign << 15) | (newexp << 10) | (significand >> 13));
         }
     }
     return fp16;
